Rename createMutator.ts
